Add a testthat load_all does update the collate order when it should
Since roxygen2::update_collate is supposed to have a permanent side effect if includes are found (viz updating the package's DESCRIPTION file) and is now called by load_all, it is necessary to make a temporary copy of the package being used for this test, otherwise future runs of the testthat will not actually be doing what is intended.

Reminder : other devtools testthats should probably take temporary copies of their templates as well if they use load_all, either directly or indirectly.