Ensure roxygen2::update_collate is up to date (>=4.1.0)
There is a bit of fiddling since there is currently (R3.1.2) a bug in
loadNamespace (and hence in requireNamespace as well).  The code here
uses a belts and brace approach to make sure the check on the version of
roxygen2 is applied bother before and after that bug is fixed.
Obviouslly, after R is fixed this code could be tidied up!