Remove S4 classes unconditionally when unloading
Any S4 classes defined by a package need to always be removed when
unloaded, not just when they were originally loaded by `load_all()`.
This prevents errors when loading a package with `library()` or
`loadNamespace()` then trying to load it with `load_all()`, due to
having old class definitions in subclasses and class unions.

Fixes #46