Fix test warnings and errors on windows
The error was caused by normalizePath having a different default for
winslash on windows. The warnings were due to devtools using pkgbuild
(so it could not be unloaded); fixed by changing the example to use
bitops. Also a warning about unloading the compiler package, which we
now test for explicitly.

Fixes #65