Ensure we only remove S4 classes from the package being unloaded
If two S4 classes A and B are defined in **pkgA**, with class A containing
class B.

    setClass("A")
    setClass("B", contains = "A")

and **pkgB** makes a new class C which contains B

    #' importClassesFrom pkgA B
    setClass("C", contains = "B")

Then `getClasses(asNamespace("pkgB"))` will return A, B, and C even
though only A should be removed.

Therefore we modify remove_s4_classes to exit early if the class is not
defined in the package being unloaded.

Fixes #75