Warn if there are conflicts between exported functions and the globalenv
Often when developing you will accidentally source a file or define a
function in the console which then masks the function exported by
`load_all()`, which can cause confusion as to why your changed in the
source file are not being used until you determine the issue.

load_all now issues an informative warning in these cases to make it
clear there is an issue and how to remedy it.

Fixes #112